:sparkles: Moviepilot
随便用用先吧..后面写 mini app 的

+++ bot-service/src/main/java/com/ocd/service/mysql/UserService.kt
@@ -18,7 +18,7 @@ class UserService {
     @Autowired
     lateinit var userMapper: UserMapper
 
-    fun createUser(user: User): Int {
+    fun createUser(user: User): Long {
         userMapper.insert(user)
         return user.id
     }
